# apps/service/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 云上 API
  version: 0.1.0
  description: |
    云上三端公共接口契约。当前先覆盖统一响应、认证、健康检查、
    平台用户列表和小程序当前会话；其他既有接口在后续修改时逐步纳入。
servers:
  - url: /
...
paths:
...
  /api/admin/platform/users:
    get:
      tags: [PlatformUsers]
      operationId: listPlatformUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: nickname
          in: query
          schema:
            type: string
        - name: phone
          in: query
          schema:
            type: string
        - name: tenantId
          in: query
          schema:
            $ref: "#/components/schemas/Identifier"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/EntityStatus"
      responses:
        "200":
          description: 平台用户分页结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformUserListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 登录状态无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 权限或租户状态不允许访问
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    Identifier:
      type: string
      pattern: "^[1-9][0-9]*$"
    EntityStatus:
      type: string
      enum: [enabled, disabled]
    ErrorResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: "null"
...
    SuccessResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          const: 0
        message:
          type: string
          const: 成功
        data: {}
...
    PlatformUser:
      type: object
      required: [id, nickname, avatarUrl, phone, status, tenantCount, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Identifier"
        nickname:
          type: [string, "null"]
        avatarUrl:
          type: [string, "null"]
        phone:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/EntityStatus"
        tenantCount:
          type: integer
          minimum: 0
        createdAt:
          type: string
    PlatformUserPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlatformUser"
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
    PlatformUserListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PlatformUserPage"
...